fix(#1767): P4 — split sync schema tests out of the asyncio-marked router test module
tests/identity/test_users_router.py carries a blanket `pytestmark =
pytest.mark.asyncio` for its router tests, which was also incorrectly
applying to TestUserSchemas' four synchronous Pydantic schema tests,
producing a PytestWarning on every tooling/backend CI run.

A class-level `pytestmark` override does NOT suppress a module-level mark
(verified empirically: pytest unions marks across module/class/function
scope rather than letting a narrower scope cancel a broader one) -- the
only correct fix is keeping sync tests out of an asyncio-marked module.
Moved TestUserSchemas to its own test_user_schemas.py with no module-level
mark; router file's UserCreate import dropped (no longer used there).
